requirements.txt
caproto == 1.0.1
p4p == 4.1.5
confluent_kafka == 2.1.0
graypy == 2.1.0
graphyte == 1.7.1
numpy == 1.19
ConfigArgParse == 1.5.3
ess-streaming-data-types == 0.21.1
tomli==2.0.1
